extract project avatar
