+++ src/components/avatars/project-avatar.js
